split into billing project and quote management db tests
